extract disk wiping from disk partitioning
